# Enable anisotropic filtering for 3D materials by default