… modules/gltf/structures/gltf_texture_sampler.h