This results in better appearance for textures viewed from oblique angles. Many other engines now enable it by default.

The anisotropic filtering quality can be configured as usual in the Project Settings. The default is 4× anisotropy.

## Preview

*I used [this Sponza demo](https://github.com/Calinou/godot-sponza-4) for the screenshots.*

### Before



### After