@Calinou 2020-07-05
@Logharaa reduz said on IRC he'd prefer this to be implemented in a more fine-grained fashion.