@reduz 2020-10-19
Mobile cant deal with this as well, and it needs to be tested on lower end devices too, so lets wait a bit on this one