@Calinou 2020-11-07
I amended the pull request to use 8× anisotropic filtering desktop and 1× on mobile (which effectively disables it there).