@reduz 2022-07-28
This needs testing on low end desktop devices to ensure it makes a difference in performance or not.