@fire 2024-11-06
@Calinou Do you think we can cover enough devices to turn this on by default?