@fire 2024-11-06
I salvaged the PR, but I'll let it simmer in my branch of Godot Engine for a bit until I submit it.

https://github.com/V-Sekai/godot/tree/vsk-default-enable-anisotropic-filtering-4.4