@Calinou 2024-11-07
Rebased and tested again (using @fire's work as a base), it works as expected.

I've also modified the glTF importer to handle this for material import (remember to reimport 3D scenes for this change to apply, e.g. by removing the `.godot/` folder).