@Calinou 2024-11-28
I did some tests of https://github.com/Calinou/godot-sponza-4 on my laptop (on AC in balanced mode):

<details open>

<summary>PC specifications</summary>

- **CPU:** Intel Core i9-13900H
- **GPU:** Intel Iris Xe / NVIDIA GeForce RTX 4060 Laptop GPU
- **RAM:** 48 GB (16 + 32 GB DDR4-3200 C22)
- **SSD:** Solidigm P44 Pro 2 TB
- **OS:** Windows 11 23H2
</details>

Type | `master` | This PR
-|-|-
Intel, Mobile, 1152x648 | 296 FPS (3.38 mspf) | 283 FPS (3.54 mspf)
Intel, Mobile, 2560x1600 | 90 FPS (11.11 mspf) | 88 FPS (11.36 mspf)
Intel, Compatibility, 1152x648 | 490 FPS (2.04 mspf) | 492 FPS (2.03 mspf)
Intel, Compatibility, 2560x1600 | 144 FPS (6.94 mspf) | 140 FPS (7.14 mspf)
NVIDIA, Mobile, 1152x648 | 1510 FPS (0.66 mspf) | 1487 FPS (0.67 mspf)
NVIDIA, Mobile, 2560x1600 | 611 FPS (1.64 mspf) | 599 FPS (1.67 mspf)
NVIDIA, Compatibility, 1152x648 | 1695 FPS (0.59 mspf) | 1727 FPS (0.58 mspf)
NVIDIA, Compatibility, 2560x1600 | 798 FPS (1.25 mspf) | 797 FPS (1.25 mspf)

The performance difference is very small on average, especially compared to the significant gain in quality. Limiting ourselves to 8x probably did the heavy lifting here, since it's basically twice as cheap as 16x with only a slight loss in visual quality. (Console versions of games often limit themselves to 8x AF for a similar reason, since CPU and GPU memory is shared.)

Note that this is a scene with no post-processing effects applied, so scenes with post-processing will spend a greater time in areas other than anisotropic texture sampling (therefore reducing the % impact of anisotropic filtering).