@Zireael07 2024-11-28
What about Android devices? They're the ones which are really low-end for Godot...