@Calinou 2024-11-28
> What about Android devices? They're the ones which are really low-end for Godot...

I've tested 8x AF on my Samsung Galaxy Z Fold4 and couldn't spot a performance difference on its main screen (it always ran at a locked 120 FPS, also on Truck Town).

However, this PR doesn't enable anisotropic filtering by default on mobile (since the `.mobile` feature tag override is used to set it to 1x, which is the same as disabled). We could revisit this in a future PR - I expect 2x AF to perform well on all mobile devices nowadays, maybe even 4x.