@clayjohn 2024-11-28
> However, this PR doesn't enable anisotropic filtering by default on mobile (since the .mobile feature tag override is used to set it to 1x, which is the same as disabled). We could revisit this in a future PR - I expect 2x AF to perform well on all mobile devices nowadays, maybe even 4x.

The performance impact of anisotropic textures doesn't work like that. There is no inherent performance penalty from using it (except for on very old devices), the performance cost comes from the additional bandwidth needed to sample it. Mobile devices are often bottlenecked by reading from textures, so multiplying the cost of texture reads is prohibitive. Desktop devices are almost never bottlenecked by texture reads, so its a free quality improvement. Remember, taking more time for something that isn't the bottleneck is essentially free (until you optimize your existing bottleneck). 

Low end integrated devices struggle more with texture reads than dedicated GPUs do. So we need to test there to assess the realistic impact. Testing on a high end device or a dedicated GPU won't give you any useful information unfortunately