@clayjohn 2025-04-14
I was doing some performance profiling on macOS last week of the TPS demo and I found that anisotropic textures were a significant cause of high bandwidth in that demo. TBDR GPUs tend to get bandwidth limited fast, so increasing bandwidth is a big concern. It would be nice if this could be scoped to non-TBDR GPUs. I'm not sure if there is an easy way for us to do that.

Unfortunately this means that this PR will likely hurt performance on any ARM-based laptops as well as all mobile devices. It might be a tradeoff worth making, but I'm not sure if it is