@fire 2025-04-14
We discussed a long time ago where system where known incompatible systems degrade to a more performant setting. Unknown systems assume to work better.