@Calinou 2025-04-15
> TBDR GPUs tend to get bandwidth limited fast, so increasing bandwidth is a big concern. It would be nice if this could be scoped to non-TBDR GPUs. I'm not sure if there is an easy way for us to do that.

If we had a way to use multiple feature tags at once in the Project Settings, we could use `.macos.arm` as an override to disable anisotropic filtering like we currently do for `.mobile`.