@v1993 2025-06-10
From my perspective this should be a pretty easy merge in its current state, as this gives a noticeable quality boost on default settings and primarily *enables* proper control over anisotropic filtering (via the existing project setting), meaning that the feature isn't forced on anyone and can be easily opted out of if it turns out to cause problems for a project. I don't think there's really a need for more override options, at least for the initial merge - on desktop anisotropic filtering is traditionally exposed as a setting to end users anyways, which those with a TBDR GPU could turn down manually if they get bottlenecked by it.